+++ tests/acceptance/api_test.go
@@ -615,7 +615,7 @@ func TestApi(t *testing.T) {
 					// Get nested json in response json
 					err := data["error"].(map[string]interface{})
 					assert.Equal(r, "ClientError", err["code"].(string))
-					assert.Equal(r, "Failed to Parse Request Body: {}",
+					assert.Equal(r, "invalid request parameters",
 						err["message"].(string))
 				},
 			})
@@ -638,7 +638,7 @@ func TestApi(t *testing.T) {
 				json:   body,
 				f: func(r *testutil.R, apiResp *apiResponse) {
 					// Verify response code
-					assert.Equal(r, http.StatusBadRequest, apiResp.StatusCode)
+					assert.Equal(r, http.StatusNotFound, apiResp.StatusCode)
 
 					// Parse response json
 					data := parseResponseJSON(t, apiResp)
@@ -646,8 +646,8 @@ func TestApi(t *testing.T) {
 					// Verify error response json
 					// Get nested json in response json
 					err := data["error"].(map[string]interface{})
-					assert.Equal(r, "ClientError", err["code"].(string))
-					assert.Equal(r, fmt.Sprintf("No leases found for Principal %q and Account ID %q", principalID, acctID),
+					assert.Equal(r, "NotFoundError", err["code"].(string))
+					assert.Equal(r, fmt.Sprintf("lease \"with Principal ID %s and Account ID %s\" not found", principalID, acctID),
 						err["message"].(string))
 				},
 			})
@@ -692,7 +692,7 @@ func TestApi(t *testing.T) {
 				json:   body,
 				f: func(r *testutil.R, apiResp *apiResponse) {
 					// Verify response code
-					assert.Equal(r, http.StatusBadRequest, apiResp.StatusCode)
+					assert.Equal(r, http.StatusNotFound, apiResp.StatusCode)
 
 					// Parse response json
 					data := parseResponseJSON(t, apiResp)
@@ -700,8 +700,8 @@ func TestApi(t *testing.T) {
 					// Verify error response json
 					// Get nested json in response json
 					errResp := data["error"].(map[string]interface{})
-					assert.Equal(r, "ClientError", errResp["code"].(string))
-					assert.Equal(r, fmt.Sprintf("No leases found for Principal %q and Account ID %q", principalID, wrongAcctID),
+					assert.Equal(r, "NotFoundError", errResp["code"].(string))
+					assert.Equal(r, fmt.Sprintf("lease \"with Principal ID %s and Account ID %s\" not found", principalID, wrongAcctID),
 						errResp["message"].(string))
 				},
 			})
@@ -745,7 +745,7 @@ func TestApi(t *testing.T) {
 				json:   body,
 				f: func(r *testutil.R, apiResp *apiResponse) {
 					// Verify response code
-					assert.Equal(r, http.StatusBadRequest, apiResp.StatusCode)
+					assert.Equal(r, http.StatusConflict, apiResp.StatusCode)
 
 					// Parse response json
 					data := parseResponseJSON(t, apiResp)
@@ -753,8 +753,8 @@ func TestApi(t *testing.T) {
 					// Verify error response json
 					// Get nested json in response json
 					errResp := data["error"].(map[string]interface{})
-					assert.Equal(r, "ClientError", errResp["code"].(string))
-					assert.Regexp(t, "Lease is not active for .*$", errResp["message"].(string))
+					assert.Equal(r, "ConflictError", errResp["code"].(string))
+					assert.Regexp(t, "leaseStatus: must be active lease", errResp["message"].(string))
 				},
 			})
 
@@ -961,15 +961,21 @@ func TestApi(t *testing.T) {
 					})
 
 					// Check the lease is decommissioned
-					// (since we dont' yet have a GET /leases endpoint
-					lease, err := dbSvc.GetLease(accountID, "test-user")
-					require.Nil(t, err)
-					require.Equal(t, db.Inactive, lease.LeaseStatus)
+					resp := apiRequest(t, &apiRequestInput{
+						method: "GET",
+						url:    apiURL + fmt.Sprintf("/leases?principalId=test-user&accountId=%s", accountID),
+						json:   nil,
+					})
+
+					results := parseResponseArrayJSON(t, resp)
+					assert.Equal(t, 200, resp.StatusCode)
+					assert.Equal(t, 1, len(results), "one lease should be returned")
+					assert.Equal(t, "Inactive", results[0]["leaseStatus"])
 
 					t.Run("STEP: Recreate lease against same account", func(t *testing.T) {
 						// Account is being reset, so it's not marked as "Ready".
 						// Update the DB to be ready, so we can create a lease
-						_, err := dbSvc.TransitionAccountStatus(accountID, db.NotReady, db.Ready)
+						_, err := dbSvc.TransitionAccountStatus(accountID, db.Leased, db.Ready)
 						require.Nil(t, err)
 
 						// Request a lease
@@ -1298,7 +1304,7 @@ func TestApi(t *testing.T) {
 			testStartDate := time.Date(currentDate.Year(), currentDate.Month(), currentDate.Day(), 0, 0, 0, 0, time.UTC)
 			testEndDate := time.Date(currentDate.Year(), currentDate.Month(), currentDate.Day(), 23, 59, 59, 59, time.UTC)
 			var testPrincipalID = "TestUser1"
-			var testAccount = "TestAccount1"
+			var testAccount = "123456789012"
 
 			t.Run("Should be able to get usage by start date and end date", func(t *testing.T) {
 				queryString := fmt.Sprintf("/usage?startDate=%d&endDate=%d", testStartDate.Unix(), testEndDate.Unix())
@@ -1322,7 +1328,7 @@ func TestApi(t *testing.T) {
 					if data[0] != nil {
 						usageJSON := data[0]
 						assert.Equal(r, "TestUser1", usageJSON["principalId"].(string))
-						assert.Equal(r, "TestAccount1", usageJSON["accountId"].(string))
+						assert.Equal(r, "123456789012", usageJSON["accountId"].(string))
 						assert.Equal(r, 2000.00, usageJSON["costAmount"].(float64))
 					}
 				})
@@ -1350,7 +1356,7 @@ func TestApi(t *testing.T) {
 					if data[0] != nil {
 						usageJSON := data[0]
 						assert.Equal(r, "TestUser1", usageJSON["principalId"].(string))
-						assert.Equal(r, "TestAccount1", usageJSON["accountId"].(string))
+						assert.Equal(r, "123456789012", usageJSON["accountId"].(string))
 						assert.Equal(r, 2000.00, usageJSON["costAmount"].(float64))
 					}
 				})
@@ -1378,7 +1384,7 @@ func TestApi(t *testing.T) {
 					if data[0] != nil {
 						usageJSON := data[0]
 						assert.Equal(r, "TestUser1", usageJSON["principalId"].(string))
-						assert.Equal(r, "TestAccount1", usageJSON["accountId"].(string))
+						assert.Equal(r, "123456789012", usageJSON["accountId"].(string))
 						assert.Equal(r, 2000.00, usageJSON["costAmount"].(float64))
 					}
 				})
@@ -1644,7 +1650,7 @@ func TestApi(t *testing.T) {
 			})
 
 			results := parseResponseArrayJSON(t, resp)
-			assert.Equal(t, 0, len(results), "only one lease should be returned")
+			assert.Equal(t, 0, len(results), "no lease should be returned")
 		})
 
 		t.Run("When there is a limit parameter", func(t *testing.T) {
@@ -2266,20 +2272,23 @@ func createUsage(t *testing.T, apiURL string, usageSvc usage.Service) {
 	timeToLive := startDate.AddDate(0, 0, ttl)
 
 	var testPrincipalID = "TestUser1"
-	var testAccountID = "TestAccount1"
+	var testAccountID = "123456789012"
 
 	for i := 1; i <= 5; i++ {
 
-		input := usage.Usage{
-			PrincipalID:  testPrincipalID,
-			AccountID:    testAccountID,
-			StartDate:    startDate.Unix(),
-			EndDate:      endDate.Unix(),
-			CostAmount:   2000.00,
-			CostCurrency: "USD",
-			TimeToLive:   timeToLive.Unix(),
-		}
-		err := usageSvc.PutUsage(input)
+		input, err := usage.NewUsage(
+			usage.NewUsageInput{
+				PrincipalID:  testPrincipalID,
+				AccountID:    testAccountID,
+				StartDate:    startDate.Unix(),
+				EndDate:      endDate.Unix(),
+				CostAmount:   2000.00,
+				CostCurrency: "USD",
+				TimeToLive:   timeToLive.Unix(),
+			},
+		)
+		require.Nil(t, err)
+		err = usageSvc.PutUsage(*input)
 		require.Nil(t, err)
 
 		usageEndDate = endDate